requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.2
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.5
botocore==1.35.5
cachetools==5.5.0
certifi==2024.7.4
cffi==1.15.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
daphne==4.1.2
db==0.1.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1
django-channels==0.7.0
django-cors-headers==4.4.0
django-heroku==0.3.1
django-livereload==1.7
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
extensions==0.4
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.66.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.133
MarkupSafe==3.0.1
matplotlib==3.9.2
multidict==6.1.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pipenv==2024.1.0
platformdirs==4.2.2
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.2
service-identity==24.1.0
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==8.5.0
tqdm==4.66.5
treelib==1.7.0
treenode==0.1.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
waitress==3.0.0
websockets==10.4
Werkzeug==3.0.4
whitenoise==6.7.0
WTForms==3.1.2
yarl==1.14.0
zipp==3.20.0
zope.interface==7.0.1
